chore: created/moved helpers for auth and cookie routes
