+++ components/auth/buttons/GoogleSSO/Button.tsx
@@ -45,7 +45,6 @@ export default function SSOButton({
       });
       router.refresh();
     } catch (error) {
-      console.info("Error in handleSuccess");
       handleError();
       setIsLoading(false);
     }
